alpha 16.6 not tested
- quest panel fixes
- change labour system -  now it is labour = workers_cf * factory_cf / work_complexity_cf
- workbuildings and worksites are now ILabourable
- try to fix farm system. plants was not updating in previous version (maybe cause of nature changes)

+++ Scripts/Nature/Plant.cs
@@ -95,9 +95,9 @@ public abstract class Plant : Structure {
     {
         return PlantType.Abstract;
     }
-    virtual public int GetPlantComplexity()
+    virtual public float GetPlantComplexity()
     {
-        return 1;
+        return 1f;
     }
     //
